faq-chatbot/requirements.txt
streamlit
langchain
langchain-google-genai
faiss-cpu
python-dotenv
langchain_community